Portal Gun
Enabled with FLAG_SYS_PORTAL_GUN_GET.
L/R fires an orange/blue portal respectively in the direction the player
is facing. The portal sticks to the first impassible tile it reaches.
Walking into one portal (from the direction it was fired) causes the
player to warp out of the other.